update to API 4 (#6)
Made adjustments to get the plugin compiling/running on Dalamud API 4.

I've also set the references to %appdata%\xivlauncher\addon\Hooks\dev as you want to target that folder for the current version of Dalamud.

DISCLAIMER: I don't actually use this plugin. Please doublecheck that it properly works! I think it does? It loaded and the bar showed and was filling up, but I don't know what/how  to test it's working as intended.